separated forms into components
